% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_to_vec.R
\name{column_to_vec}
\alias{column_to_vec}
\title{Column to vector
Takes a column of strings and puts them into an R-vector.}
\usage{
column_to_vec()
}
\value{
The vector.
# Example: Mark A, B, C  and press CTRL + C:
# A
# B
# C
}
\description{
This function is used to convert a column of strings into an R-vector without inserting commas manually.
}
