% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.choose.R
\name{file.choose}
\alias{file.choose}
\title{file.choose
Inserts the selected string into the clipboard}
\usage{
file.choose()
}
\value{
Prints the selected string

file.choose()
}
\description{
This function copies the selected string into the clipboard, to avoid using the mouse.
}
