% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldply_template.R
\name{ldply_template}
\alias{ldply_template}
\title{ldply_template
Inserts an ldply-template into the clipboard}
\usage{
ldply_template(split_variable = NULL)
}
\arguments{
\item{split_variable}{The variable to split the data frame by.}
}
\value{
Nothing
}
\description{
Inserts a skeleton ldply-template into the clipboard
}
