% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_packages.R
\name{library_packages}
\alias{library_packages}
\title{library_packages
Installs and loads my favourite packages}
\usage{
library_packages(package_vec = NULL)
}
\value{
Nothing

library_packages()
}
\description{
Installs and loads my favourite packages
}
