% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_temptable_template.R
\name{sql_temptable_template}
\alias{sql_temptable_template}
\title{sql-temptable-template
Inserts an sql-temptable-template into the clipboard}
\usage{
sql_temptable_template(temptable_name = NULL)
}
\arguments{
\item{temptable_name}{If a temptable_name is provided, it is inserted into the drop-table and create-temptable-positions in the template.}
}
\value{
Nothing
}
\description{
Inserts a sql-temptable-template into the clipboard
}
