% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_SafetyReportID_to_UMCReportID.R
\name{translate_SafetyReportID_to_UMCReportID}
\alias{translate_SafetyReportID_to_UMCReportID}
\title{translate_SafetyReportID_to_UMCReportID}
\usage{
translate_SafetyReportID_to_UMCReportID(
  safety_report_ids,
  database = "UMCReportWeekly"
)
}
\value{
The UMCReportID as a character vector.
}
\description{
Querys UMCDB06, UMCReportWeekly (by default) and translates SafetyReportIDs to UMCReportID using the Report-table.
}
