% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_UMCReportID_to_SafetyReportID.R
\name{translate_UMCReportID_to_SafetyReportID}
\alias{translate_UMCReportID_to_SafetyReportID}
\title{translate_UMCReportID_to_SafetyReportID}
\usage{
translate_UMCReportID_to_SafetyReportID(
  umc_report_ids,
  database = "UMCReportWeekly"
)
}
\value{
The SafetyReportIDs as a character vector.
}
\description{
Querys UMCDB06, UMCReportWeekly (by default) and translates UMCReportIDs to SafetyReportIDs.
}
