% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_UMCReportID.R
\name{update_UMCReportID}
\alias{update_UMCReportID}
\title{update_UMCReportID}
\usage{
update_UMCReportID(report_ids_to_update, database = "UMCReportWeekly")
}
\value{
The updated UMCReportID as a character vector.
}
\description{
Querys UMCDB06, UMCReportWeekly (by default) and updates any passed UMCReportID to the latest mapped reportID in the follow-up-table.
}
