% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_packages.R
\name{library_packages}
\alias{library_packages}
\title{library_packages
Installs and loads my favourite packages}
\usage{
library_packages(package_vec = NULL)
}
\arguments{
\item{package_vec}{Provide the function with custom packages to install and load. Default is "plyr", "dplyr", "magrittr", "data.table", "stringr" and "lubridate".}
}
\value{
Nothing

library_packages()
}
\description{
Installs and loads my favourite packages
}
